refactor: simplify DialogMap component data flow and state management
